+++ lyzr/base/prompt.py
@@ -1,4 +1,3 @@
-import json
 from typing import Literal
 from lyzr.base.prompt_texts import PROMPT_TEXTS
 from lyzr.base.base import ChatMessage, UserMessage, SystemMessage, MessageRole
